refactor!: compat→match 전면 리네임 + 안티패턴 리뷰 반영 (P0 2건·P1 5건)
리네임 (breaking, 프리릴리스라 무마이그레이션):
- matchSheet/formatMatchSheet/judgeMatch, Match* 타입, src/match/, schemaVersion match-sheet-v2
- 렌더 출력(GLM 입력 텍스트)은 바이트 단위 불변 — 5쌍×3포맷 diff로 검증(재라벨 불필요)

안티패턴 수정 (코덱스 리뷰 → 직접 재현 검증분만):
- P0: catalog slotIndex Int8→Int16 (hours 128개 이상에서 음수 래핑 재현·수정)
- P0: setSolarConfig가 undefined 값 패치를 무시 + boolean 런타임 검증
  (JS에서 전역 설정 오염 → bazi()가 빈 시주 반환하던 것 차단)
- P1: 미구현 timezone 필드 제거 (조용히 무시되는 공개 옵션)
- P1: formatter가 표시 라벨 파싱 대신 id→한자 상수맵 사용 (라벨 문구 변경이
  GLM 입력을 못 흔들게; 미등록 id만 괄호 파싱 폴백)
- P1: 코어 Subject/SheetSubject 도입 — analyze는 Subject, baziSheet는
  SheetSubject(4기둥·성별 필수, 타입=런타임 전제조건), bazi()에 hour 지정 시
  시주 non-null 오버로드 추가. 궁합 타입 누출 해소(MatchSubject=Subject 별칭)
- P1: tsconfig.test.json + typecheck:test — 테스트도 strict 타입체크
  (드러난 타입 오류 전부 수정: gender 여분 키, JSON import 속성, never 인덱싱 등)

검증: build·typecheck·typecheck:test 클린, 680 테스트 그린,
slotIndex·config 오염·baziSheet 전제조건 재현 프로브 전부 통과.

+++ src/catalog.ts
@@ -61,7 +61,8 @@ export function catalog(
   const monthsArr  = new Int8Array(N)
   const daysArr    = new Int8Array(N)
   const hoursArr   = new Int8Array(N)
-  const slotIndex  = new Int8Array(N)
+  // Int16 — hours 인자에 중복 포함 128개 이상이 와도 인덱스가 음수로 래핑되지 않게.
+  const slotIndex  = new Int16Array(N)
   const stemsArr   = new Int8Array(N * 4)
   const branchesArr = new Int8Array(N * 4)
